+++ packages/terra-notification-dialog/src/terra-dev-site/doc/example/CompleteNotificationDialog.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import Button from 'terra-button';
 import NotificationDialog, { NotificationDialogVariants } from 'terra-notification-dialog';
 
-const clickOK = () => {
-  console.log('You clicked OK'); // eslint-disable-line no-console
+const clickConfirm = () => {
+  console.log('You clicked confirm'); // eslint-disable-line no-console
 };
 
 class CompleteNotificationDialog extends React.Component {
@@ -33,15 +33,17 @@ class CompleteNotificationDialog extends React.Component {
           variant={NotificationDialogVariants.ALERT}
           isOpen={this.state.isOpen}
           title="Make sure that the title relates directly to the choices."
-          message="The Main Instruction is text used to provide more detail or define terminology. Don’t repeat the title verbatim."
-          primaryAction={{
-            text: 'OK',
-            onClick: clickOK,
+          startMessage="The Main Instruction is text used to provide more detail or define terminology. Don’t repeat the title verbatim."
+          acceptAction={{
+            text: 'Confirm',
+            onClick: clickConfirm,
           }}
-          secondaryAction={{
+          rejectAction={{
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
+          buttonOrder="acceptFirst"
+          emphasizedAction="accept"
         />
         <Button text="Trigger NotificationDialog" onClick={this.handleOpenModal} />
       </div>