+++ packages/terra-notification-dialog/src/terra-dev-site/doc/example/NotificationDialogWithLongText.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import Button from 'terra-button';
 import NotificationDialog, { NotificationDialogVariants } from 'terra-notification-dialog';
 
-const clickOK = () => {
-  console.log('You clicked OK'); // eslint-disable-line no-console
+const clickConfirm = () => {
+  console.log('You clicked confirm'); // eslint-disable-line no-console
 };
 
 class NotificationDialogWithLongText extends React.Component {
@@ -112,15 +112,17 @@ class NotificationDialogWithLongText extends React.Component {
           variant={NotificationDialogVariants.WARNING}
           isOpen={this.state.isOpen}
           title="Make sure that the title relates directly to the choices."
-          message={message}
-          primaryAction={{
-            text: 'OK',
-            onClick: clickOK,
+          startMessage={message}
+          acceptAction={{
+            text: 'Confirm',
+            onClick: clickConfirm,
           }}
-          secondaryAction={{
+          rejectAction={{
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
+          buttonOrder="acceptFirst"
+          emphasizedAction="accept"
         />
         <Button text="Trigger NotificationDialog" onClick={this.handleOpenModal} />
       </div>