+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/ContentNotificationDialog.test.jsx
@@ -6,7 +6,7 @@ const clickConfirm = () => {
   alert('You clicked confirm'); // eslint-disable-line no-alert
 };
 
-class CompleteNotificationDialogWithLongMessage extends React.Component {
+class ContentNotificationDialog extends React.Component {
   constructor() {
     super();
 
@@ -30,10 +30,17 @@ class CompleteNotificationDialogWithLongMessage extends React.Component {
     return (
       <div>
         <NotificationDialog
-          variant={NotificationDialogVariants.SUCCESS}
+          variant={NotificationDialogVariants.ALERT}
           isOpen={this.state.isOpen}
-          title="This is the titleThis is the titleThis is the titleThis is the titleThis is the titleThis is the titleThis is the titleThis is the title"
-          startMessage="This is the messageThis is the messageThis is the messageThis is the messageThis is the messageThis is the messageThis is the messageThis is the messageThis is the messageThis is the message"
+          title="Make sure that the title relates directly to the choices."
+          startMessage="The Main Instruction is text used to provide more detail or define terminology. Don’t repeat the title verbatim."
+          content={(
+            <ul>
+              <li>Item 1</li>
+              <li>Item 2</li>
+            </ul>
+          )}
+          endMessage="The Main Instruction is text used to provide more detail or define terminology. Don’t repeat the title verbatim."
           acceptAction={{
             text: 'Confirm',
             onClick: clickConfirm,
@@ -51,4 +58,4 @@ class CompleteNotificationDialogWithLongMessage extends React.Component {
   }
 }
 
-export default CompleteNotificationDialogWithLongMessage;
+export default ContentNotificationDialog;