+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/NotificationDialogMinimalProps.test.jsx
@@ -27,10 +27,12 @@ class NotificationDialogMinimalProps extends React.Component {
       <div>
         <NotificationDialog
           isOpen={this.state.isOpen}
-          primaryAction={{
+          acceptAction={{
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
+          buttonOrder="acceptFirst"
+          emphasizedAction="accept"
         />
         <Button id="trigger-notification-dialog" text="Trigger NotificationDialog" onClick={this.handleOpenModal} />
       </div>