+++ packages/terra-notification-dialog/tests/wdio/notification-dialog-spec.js
@@ -103,6 +103,34 @@ Terra.describeViewports('notification-dialog', ['tiny', 'medium', 'large'], () =
       Terra.validates.element({ selector });
     });
   });
+
+  describe('Notification-dialog with Content', () => {
+    it('Notification-dialog with Content', () => {
+      browser.url('/#/raw/tests/terra-notification-dialog/notification-dialog/content-notification-dialog');
+      browser.click('#trigger-notification-dialog');
+      Terra.validates.element({ selector });
+    });
+  });
+
+  describe('Notification-dialog with emphasized reject action', () => {
+    it('Notification-dialog with emphasized reject action', () => {
+      browser.url('/#/raw/tests/terra-notification-dialog/notification-dialog/emphasized-reject-notification-dialog');
+      browser.click('#trigger-notification-dialog');
+      Terra.validates.element({ selector });
+    });
+  });
+});
+
+Terra.describeViewports('notification-dialog', ['tiny', 'medium', 'large'], () => {
+  afterEach(() => browser.click('[class*="NotificationDialog-module__actions"] button:first-child'));
+
+  describe('Notification-dialog with Reversed Actions', () => {
+    it('Notification-dialog with Reversed Actions', () => {
+      browser.url('/#/raw/tests/terra-notification-dialog/notification-dialog/reversed-action-notification-dialog');
+      browser.click('#trigger-notification-dialog');
+      Terra.validates.element({ selector });
+    });
+  });
 });
 
 Terra.describeViewports('NotificationDialog with additional focus trap sources within a modal manager', ['medium'], () => {